fixed band photo

+++ src/bands/bands.service.ts
@@ -3,6 +3,8 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Band } from "./entities/band.entity";
 import { CreateBandDto } from "./dto/create-band.dto";
+import { join } from "path";
+import * as fs from 'fs';
 
 @Injectable()
 export class BandsService {
@@ -11,11 +13,20 @@ export class BandsService {
     private readonly bandRepository: Repository<Band>,
   ) {}
 
-  async create (CreateBandDto: CreateBandDto, imageUrl?: string): Promise<Band> {
-    const band = this.bandRepository.create({
-      ...CreateBandDto,
-      imageUrl,
-    });
+  // async create (CreateBandDto: CreateBandDto, imageUrl?: string): Promise<Band> {
+  //   const band = this.bandRepository.create({
+  //     ...CreateBandDto,
+  //     imageUrl,
+  //   });
+  //   return await this.bandRepository.save(band);
+  // }
+  async create(data: any) {
+    const band = new Band();
+    band.name = data.name;
+    band.genre = data.genre;
+    band.description = data.description;
+    band.imageUrl = data.imageUrl;
+
     return await this.bandRepository.save(band);
   }
 
@@ -38,7 +49,23 @@ export class BandsService {
   }
 
   async remove(id: number): Promise<void> {
-    const band = await this.findOne(id);
+    const band = await this.bandRepository.findOne({ where: { id } });
+
+    if (!band) {
+      throw new NotFoundException('Band not found');
+    }
+
+    if (band.imageUrl) {
+      const filePath = join(process.cwd(), 'public', band.imageUrl);
+
+      fs.unlink(filePath, (err) => {
+        if (err) {
+          console.error('Failed to delete image')
+        } else {
+          console.log("Successfully deleted image")
+        }
+      })
+    }
     await this.bandRepository.remove(band);
   }
 }
